More voice stuff, fix packet serialization
This commit fixes an important issue with some packets passing
JSONValues to the parent BasePacket class for serilization. This
behavior will actually convert the value into a JSONValue, then a
variant, and then finally back to a JSONValue. This was actually fuxing
with types and causing things like ushort's to be represented as strings
in payloads (which discord doesnt like, obviously)